Add source data for stats+classes and mysqldump
I did a lot of stuff before committing this. Most of it is kind of
nonsense, as ultimately the database is going to be the main end result.
The backup idea was probably a good call though. I went for a logical
backup because I like the idea of that.