Add abilities to class table + more csvs
So my poor SQL knowledge made this a bit of a learning curve. The magic
solution required was the JOIN idea.

This made the solution something like this:
	import class-abilities into a temporary table
	join this table to class via the name
	set the upgrades column in class to use the temp data

While trying to get it working I ended up (finally!) setting a primary
key for the class table. Not sure whether I'll end up adding an 'id'
style primary key later on as originally intended or if I'll trust the
data sources to stick with the textual names. I suppose eventually all
tables really ought to have a primary key too.