Preparing to reformat upgrade paths for import
I want each 'family' of classes to be labelled as such. To do this, I
plan to end up with a table similar to the following:

	Class name, class family
	Class name, class family
	etc.

So for instance

	Squire, Squire
	Knight, Squire
	Gallant, Squire
	...

I can't think of how to trick vim into reformatting the current
upgrade-paths.txt file like this, so I've turned to sed/awk/python.
transform-file is based off techniques in the following article which
need refining before pulling the trigger:
http://stackabuse.com/read-a-file-line-by-line-in-python/
                                          ,