Finish first draft of modifiation script + backup
Interesting.

The script has its quirks, such as inserting newlines into the output.
Of course, these are straightforward to remove in vim, but there were a
few other quirks I couldn't quite get my head around too.

For example, in the corresponding script from the cited source, using
line.strip() works fine, but if you remove the references to cnt then
line seems to come through as an array, which I don't really understand.
There are a lot of aspects of the script whose subtleties I glossed over
though so I'm hardly surprised.

Anyway, it seems to work so I'm going to test importing it and we'll
make sure the transformed data is correct.

Also add more current backup.