Add 'logic' section
Python was always likely to be the business end of this project for
making calculations etc., and I feel it's time to start on that.

With my previous experience, I couldn't think of a satisfactory way to
structure the code. I remembered about an interesting article I read
some time ago about OOP in Python, which I revisited. This turned out to
be exactly what I wanted.

As of this commit, the main python file is somewhat basic and
unfinished, but should more or less be the starting point to develop
from.

I've also boiled the project down to three main components now:
	SQL database - almost done
	Python logic - started
	Connecting the two - no idea

For now, since I'll probably be working on the train and whatnot, I'll
probably keep on with the second aspect and try to sort the third at
home. We'll see how things unfold though.

Side notes: a) the article I mention is this one:
https://jeffknupp.com/blog/2014/06/18/
improve-your-python-python-classes-and-object-oriented-programming/

b) I should probably put some of that commit message in a README