First success properly retrieving info from db
Wow, this feels incredible! I did it!

So, the init method of the Adept class finally actually does something.
Just one query to pull base element power/resist is up and running, but
it bloody well works. Once the adepts are initialised at the bottom of
the file, you can see they have the correct attributes
(isaac.venuspower_base comes back as 79, for example).

I'm not sure the way it's been done is optimal, but I can sort that out
later. For instance, I'm not sure if the for loop is strictly necessary,
or whether an intermediate 'assignment' function of some kind could be
used.

Hooray! This feels like a major milestone commit :)