Add buffs to weapons and armor
After musing on how best to do this, including consulting a more
knowledgeable friend, I decided to cheat and make a string based code to
act as buff values.

Each code word consists of 3 letters to specify what stat should be
modified, + - or x to indicate whether the stat is being added to,
subtracted from or multiplied, followed by the value to be used in said
operation.

The buffs are currently ignored by main but can be fairly easily
accounted for, as modifying the query to retrieve the buffs will be
trivial, and using string.split() to parse the values should hopefully
be quite straightforward too.

The only passive buff that now hasn't been entered is for a ring,
possibly the guardian ring.